+++ src/lib.rs
@@ -40,8 +40,18 @@ extern crate time_test;
 #[macro_use]
 mod infolog;
 
-#[cfg(feature = "asyncdb")]
+#[cfg(any(feature = "asyncdb-tokio", feature = "asyncdb-async-std"))]
 mod asyncdb;
+
+#[cfg(feature = "asyncdb-tokio")]
+mod asyncdb_tokio;
+#[cfg(feature = "asyncdb-tokio")]
+use asyncdb_tokio::{send_response, send_response_result, Message};
+
+#[cfg(feature = "asyncdb-async-std")]
+mod asyncdb_async_std;
+#[cfg(feature = "asyncdb-async-std")]
+use asyncdb_async_std::{send_response, send_response_result, Message};
 
 mod block;
 mod block_builder;
@@ -82,9 +92,10 @@ mod db_iter;
 pub mod compressor;
 pub mod env;
 
-#[cfg(feature = "asyncdb")]
-pub use asyncdb::AsyncDB;
-
+#[cfg(feature = "asyncdb-async-std")]
+pub use asyncdb_async_std::AsyncDB;
+#[cfg(feature = "asyncdb-tokio")]
+pub use asyncdb_tokio::AsyncDB;
 pub use cmp::{Cmp, DefaultCmp};
 pub use compressor::{Compressor, CompressorId};
 pub use db_impl::DB;
